requirements.txt
nextcord>=2.5.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
